# docs/swagger.yaml
basePath: /api
definitions:
  handlers.BatchReviewItem:
    properties:
      post_id:
        example: 12345
        type: integer
      status:
        enum:
        - approved
        - rejected
        example: approved
        type: string
    type: object
  handlers.BatchReviewRequest:
    properties:
      disable_rejected:
        example: true
        type: boolean
      post_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      status:
        enum:
        - approved
        - rejected
        example: approved
        type: string
      videos:
        items:
          $ref: '#/definitions/handlers.BatchReviewItem'
        type: array
    type: object
  handlers.BatchReviewResponse:
    properties:
      disabled:
        type: integer
      failed:
        type: integer
      results:
        items:
          $ref: '#/definitions/handlers.BatchReviewResult'
        type: array
      state:
        $ref: '#/definitions/handlers.ReviewState'
      success:
        type: integer
      total:
        type: integer
    type: object
  handlers.BatchReviewResult:
    properties:
      disabled:
        type: boolean
      message:
        type: string
      post_id:
        type: integer
      status:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.ReviewState:
    properties:
      recheck_all:
        type: boolean
      reviewed_ids:
        items:
          type: integer
        type: array
      statuses:
        additionalProperties:
          type: string
        type: object
    type: object
  handlers.ReviewStatusRequest:
    properties:
      post_id:
        example: 12345
        type: integer
      recheck_all:
        type: boolean
      status:
        enum:
        - approved
        - rejected
        example: approved
        type: string
    type: object
...
info:
  contact: {}
  description: A Go-Gin based API for video center management. Provides paginated
    or full video lists, batch video toggles, and batch AI/manual review results.
  title: Video Center API
  version: "1.0"
paths:
...
  /review/state:
    delete:
      description: Clear all reviewed video IDs from server-side JSON file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReviewState'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Clear review state
      tags:
      - review
    get:
      description: Returns the detailed review status and compatibility list of completed
        video IDs stored in server-side JSON file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReviewState'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get review state
      tags:
      - review
    post:
      consumes:
      - application/json
      description: Save a video's approved or rejected result, or toggle recheck mode
        for all videos without clearing their results.
      parameters:
      - description: Video review status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ReviewStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReviewState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Add reviewed video ID
      tags:
      - review
  /review/batch:
    post:
      consumes:
      - application/json
      description: Save approved or rejected results for multiple videos. Use videos
        for mixed statuses, or post_ids plus status for one status. Rejected videos
        are disabled upstream by default.
      parameters:
      - description: Batch review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BatchReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Batch save video review results
      tags:
      - review
...
schemes:
- http
- https
swagger: "2.0"